# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.RefundDecisionRequest:
    properties:
      review_note:
        maxLength: 500
        type: string
      status:
        enum:
        - approved
        - rejected
        type: string
    required:
    - status
    type: object
  requests.RefundRequestPayload:
    properties:
      reason:
        maxLength: 500
        minLength: 5
        type: string
      ticket_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - reason
    - ticket_ids
    type: object
...
  responses.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  responses.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/responses.ErrorDetail'
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  responses.RefundResponse:
    properties:
      amount_minor:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      debited_from_organizer_balance:
        type: boolean
      id:
        type: string
      order_id:
        type: string
      processed_at:
        type: string
      provider_refund_reference:
        type: string
      reason:
        type: string
      requested_by_id:
        type: string
      review_note:
        type: string
      reviewed_at:
        type: string
      reviewed_by_id:
        type: string
      status:
        type: string
      ticket_ids:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  responses.TicketResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      event_id:
        type: string
      holder_email:
        type: string
      holder_name:
        type: string
      holder_phone:
        type: string
      id:
        type: string
      issued_at:
        type: string
      occurrence_id:
        type: string
      order_id:
        type: string
      order_item_id:
        type: string
      scan_location:
        type: string
      status:
        type: string
      ticket_type_id:
        type: string
      transfer_count:
        type: integer
      updated_at:
        type: string
      used_at:
        type: string
      user_id:
        type: string
    type: object
...
  responses.TicketWithQRResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      event_id:
        type: string
      holder_email:
        type: string
      holder_name:
        type: string
      holder_phone:
        type: string
      id:
        type: string
      issued_at:
        type: string
      occurrence_id:
        type: string
      order_id:
        type: string
      order_item_id:
        type: string
      qr_payload:
        type: string
      scan_location:
        type: string
      status:
        type: string
      ticket_type_id:
        type: string
      transfer_count:
        type: integer
      updated_at:
        type: string
      used_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@1kulture.com
    name: API Support
    url: https://1kulture.com/support
  description: Enterprise Event Management System API with multi-tenant SaaS architecture
  license:
    name: Proprietary
    url: https://1kulture.com/license
  termsOfService: https://1kulture.com/terms
  title: 1Kulture API
  version: "1.0"
paths:
...
  /admin/refunds:
    get:
      consumes:
      - application/json
      description: Admin only.
      parameters:
      - description: Status
        in: query
        name: status
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Refunds
          schema:
            items:
              $ref: '#/definitions/responses.RefundResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List refunds by status (admin)
      tags:
      - admin-refunds
...
  /events/{id}/tickets:
    get:
      consumes:
      - application/json
      description: Organizer/co-organizer only.
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tickets
          schema:
            items:
              $ref: '#/definitions/responses.TicketResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List tickets sold for an event
      tags:
      - tickets
...
  /orders/{id}/refunds:
    get:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Refunds
          schema:
            items:
              $ref: '#/definitions/responses.RefundResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List refunds for an order
      tags:
      - refunds
    post:
      consumes:
      - application/json
      description: Buyer requests a refund for specific tickets.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Refund
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.RefundRequestPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Requested
          schema:
            $ref: '#/definitions/responses.RefundResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request a refund
      tags:
      - refunds
...
  /refunds/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Refund ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Refund
          schema:
            $ref: '#/definitions/responses.RefundResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a refund
      tags:
      - refunds
  /refunds/{id}/decide:
    post:
      consumes:
      - application/json
      description: Organizer/co-organizer only.
      parameters:
      - description: Refund ID
        in: path
        name: id
        required: true
        type: string
      - description: Decision
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.RefundDecisionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reviewed
          schema:
            $ref: '#/definitions/responses.RefundResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve or reject a refund
      tags:
      - refunds
...
  /tickets/{id}:
    get:
      consumes:
      - application/json
      description: Buyer or event organizer can view.
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ticket
          schema:
            $ref: '#/definitions/responses.TicketWithQRResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a ticket by ID
      tags:
      - tickets
...
  /tickets/code/{code}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Ticket code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ticket
          schema:
            $ref: '#/definitions/responses.TicketWithQRResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a ticket by code
      tags:
      - tickets
  /tickets/me:
    get:
      consumes:
      - application/json
      description: All tickets owned by the authenticated user.
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tickets
          schema:
            items:
              $ref: '#/definitions/responses.TicketResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my tickets
      tags:
      - tickets
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"